Break up modules to folders/namespaces
