+++ tests/src/Payloads/PayloadFactoryTest.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Tarcha\WebKernel\Tests\Payloads;
+
+use Tarcha\WebKernel\Tests\Test;
+
+class PayloadFactoryTest extends Test
+{
+    public function testPayloadNoContent()
+    {
+
+    }
+}
